# src/main/resources/openapi/product-openapi.yaml
openapi: "3.0.3"
info:
  title: inventory-api
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080
paths:
  /products:
    get:
      tags:
        - Product
      description: 获取所有产品
      operationId: getAllProducts
      responses:
        '200':
          description: 成功返回所有产品
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '404':
          description: 未找到任何产品
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Product
      description: 添加产品到库存
      operationId: addProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: 产品添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: 请求错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /products/{id}:
    get:
      tags:
        - Product
      description: 根据 ID 获取产品
      operationId: getProductById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
            description: 产品 ID
          required: true
      responses:
        '200':
          description: 成功获取产品
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: 未找到产品
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Product
      description: 更新产品信息
      operationId: updateProduct
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
            description: 产品 ID
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: 产品更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: 请求错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 产品不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags:
        - Product
      description: 删除指定 ID 的产品
      operationId: deleteProductById
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
            description: 产品 ID
          required: true
      responses:
        '204':
          description: 产品删除成功
        '404':
          description: 产品不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    Pageable:
      description: group Pageable query parameters
      type: object
      properties:
        page:
          type: integer
        size:
          type: integer
    Page:
      description: minimal Page response without content property
      type: object
      properties:
        number:
          type: integer
          description: page number
          example: 0
        size:
          type: integer
          description: page size
          example: 10

    ProductPage:
      description: combine Page with typed Page content
      type: object
      properties:
        totalSize:
          type: integer
        data:
          $ref: '#/components/schemas/ProductList'


    ProductList:
      type: array
      items:
        $ref: '#/components/schemas/Product'

    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 产品 ID
        name:
          type: string
          description: 产品名称
        type:
          type: integer
          format: int32
          description: 产品类型
        create_time:
          type: string
          format: date-time
          description: 产品创建时间
        price:
          type: number
          format: double
          description: 产品价格
        show:
          type: boolean
          description: 产品是否显示
        password:
          type: string
          format: password
          description: 产品密码
    Error:
      type: object
      properties:
        message:
          type: string